Initial commit
Supports adding the same set of statements to a list of files. We reuse
the WikibaseMediaInfo UI, slightly hacking it to support editing
different revisions (somewhat fragile, but if it breaks, it at least
shouldn’t edit the wrong entity, because the initial entity ID is
invalid). The OOUI glue is a bit ugly, I’ll try to improve it later.
Also, the file input via TagMultiSelectWidget isn’t great, you can’t
really paste multiple files into it – I might try to improve that (split
on |, for example), and/or add other modes (e. g. PagePile).