Use ProcessDialog instead of plain Dialog
I didn’t initially use this because I misread it as “ProgressDialog”
when I first saw it in the documentation, but it’s actually the right
thing for this use case, I think – it certainly looks nicer and gives us
“cancel” (beyond Esc, which always worked) practically for free.

Because the save now happens outside of the initialize scope, the
filesWidget and statementWidgets need to move to `this` instead of just
being plain variables.

We can also use this.close() instead of emitting an event and then
closing the dialog via the window manager; that’s not strictly related,
but let’s do it anyways in this commit.