Improve dialog height
Instead of hard-coding a value that kinda works on my main monitor,
properly adapt to the content height – I finally figured out how to do
it properly. The most important parts are to call updateSize(), because
that doesn’t happen magically, and to *not* make the PanelLayout
expanded, so that it actually has a size we can calculate. Also, we
apply rounding to avoid overly frequent size changes. (I might actually
make that even coarser in the future – perhaps start out at 300 and
increase in multiples of 300?)