Detect incompatible changes to StatementWidget
In general, this script is rather brittle and prone to break when
incompatible changes to WikibaseMediaInfo are made – but changes to
getChanges() and getRemovals() would be especially dangerous (if our
getRemovals() override no longer works, the script might start removing
existing statements from entities), and we can at least test if they’re
in the prototype where we expect them fairly easily.

The implementation is a bit more generic because I intend to add a few
more sanity checks soon.