Don’t duplicate existing statements
Download the current statements of each entity, and only add statements
if they don’t exist yet. Note that this is a very strict comparison –
for example, if a statement already exists without a qualifier, and the
new statements specify the same statement and value with some qualifier,
then a new statement with that qualifier will be added, instead of, for
example, adding the qualifier to the existing statement. I might revisit
this later, but this is already a lot better than the previous state.