Use FieldsetLayout for content
This adds some spacing that was definitely necessary, and also labels to
make it easier to understand what the files input is for (especially
given that the placeholder is not initially visible since the input
starts out at width 0px, but I hope to fix that as well later).

The labels are made into headings using some semi-nice trickery – we
have to wrap the heading around the header, not put it inside it,
because the header (<span>) or the label (<label>) only allow phrasing
content inside them.