Fix file input placeholder visibility
The file input’s actual <input> starts with a width of 0px; we need to
instruct it to resize itself once it’s embedded in the DOM.

For reasons I don’t quite understand, this needs to happen
asynchronously, though a timeout of 0ms seems to be enough; from looking
at the OOUI documentation, it looks like initialize() is actually the
wrong method to do all this setup: if we want the content to be new each
time, we should override getSetupProcess() instead. I’ll do that
separately, though.