Update files widget input size in ready process
Both “initialize” and “setup” are too early for this – “ready” is where
we can do it.

I think I misunderstood the purpose of initialize() and
getSetupProcess() in the commit message of the previous commit –
initialize() *is* the right place to set up the DOM elements, and if
someone wants the content to be new each time, they should create new
StatementsDialog instances instead of us pretending they’re somehow
reusable. (Instead, reopening the same instance should preserve the
state.)