Allow separating multiple files with pipes
This makes it much easier to use the script with multiple files: if you
have a pipe-separated list from somewhere else, you can just paste it
into the input and it will be split on Enter. (File names are titles and
cannot contain the pipe, or several other characters.) The placeholder
is updated to indicate this option with a second example file name.

This is implemented by overriding the addTagFromInput() method – if we
tried to do it in the event handler, it would be more difficult to
suppress the event for the original, unsplit tag. addTagFromInput() is
not marked as private, so hopefully this isn’t too fragile.