Automatically add selected files as tags
Previously, selecting one of the lookup suggestions would just set the
input to that value, and you’d have to press Enter again to turn it into
a tag. Let’s make it a tag directly.

There might be a better event name to emit here, I’m not sure – but at
least this doesn’t couple the two widgets too tightly together.