Add progress bar during save
Estimate the progress based on the number of statements. It’s not
entirely accurate, given that some statements may not need to be saved
if they already exist, and it doesn’t take the entity ID resolving and
entity data downloading into account yet either, but still looks pretty
nice and is hopefully useful.