Extract StatementsProgressBarWidget class
Moving the calculations inside this class keeps the action handler
simpler and focused on its real job (saving statements). I also plan to
complicate the calculation somewhat, by accounting the initial loading
of entity IDs and data as some progress (added a TODO for that) – that’s
not happening tonight, though.