Improve progress bar
First off, we don’t need to wait until the two awaits to properly enable
it – the number of entities is the same as the number of titles, we know
that before turning titles into entity IDs or loading their data.

Second, we can also estimate how many API calls will be made in total,
including the ones to load the entity IDs and data, and then update the
progress taking this into account to give a more accurate picture. (Not
all statements actually require an API call – those that are already
present are skipped – and not all API calls take the same time either,
but I think this is as good as I’m willing to implement. We don’t need
anything even more complicated. This is fine.)