Lazily spawn MultiStatements via factory
Register MultiStatements in a factory, and instead of opening it
immediately, add a link to the sidebar and open the dialog when that
link is clicked. This is much more suitable for a user script, though I
expect I’ll add an “immediately open” mode for development soon.

The window is reused when you close it and click the link again, which
isn’t really intentional, but I’ll investigate how to fix that later.