Don’t suggest already selected files
Synchronize the list of already selected titles between the FilesWidget
and the FileInputWidget, and make the FileInputWidget skip them when
creating a new lookup menu.

Note that this can leave the user with the impression that there are no
more search results – the default limit is 10, and we never continue the
query beyond that. I don’t see a good way to fix that, though: we can
bump the limit, of course (the max is 500), but even bumping it to 50
results in a noticeably slower response, so I don’t want to do it for
all requests. Perhaps I’ll improve this later.