Disallow multiple statements with the same main value
I plan to add support for updating existing statements with the same
value very soon, and in that case the behavior if the user specifies
multiple statements with the same value would be confusing and
ambiguous. Let’s disallow it for now – as soon as the user adds such a
statement, we disable the publish button, mark the conflicting
statements, and make a notice at the bottom visible. Once all duplicate
statements are removed, we enable the publish button again and hide the
notice once more.

The indicator on the duplicate statements is pretty lazy (I recently
read that “border: 1px solid red is the console.log() of CSS”), but
looks acceptable for now. I left a TODO for improving it later :)