Update existing statements
If statements with the same main value exist, add qualifiers to them
instead of adding entirely new statements. This matches QuickStatements
and is also likely to be the most useful mode.

Unfortunately, we need two new modules for this – in JS, statements
don’t have a clone or copy method, so serializing + deserializing is the
best way I know to copy one.