Add button to add files from PagePile
The “button” is the icon of the TagMultiselectWidget – I tried for a
while to make that a real button but it didn’t work out, so for now we
add a pointer cursor and click handler on the standard icon element and
call it a day. Clicking it opens a menu with only one button for now, to
load a PagePile – in theory I could add more buttons here (e. g.
directly add category members without a PetScan+PagePile detour), but
that would need more improvements for the layout because right now it
doesn’t look great if the buttons don’t have the same width. Clicking
that button opens a popup to ask for the PagePile ID and then loads the
pile and adds it to the widget.

My original vision was to add a separate input for the PagePile ID, and
load the pile only after “submit” is clicked, but I think this version
is actually much better: it gives you a sense for the scale of the pile
before you submit anything (for large piles we even emit a warning), and
if you want, you can even teak the files afterwards (though I’m not sure
how useful that is). And the PagePile UI is more out of the way this
way, instead of taking up space (beyond the icon) all the time.