Set up webpack and Babel
Instead of directly saving acdc.js as the user script, we build it using
webpack, including the Babel module, into a script that should support
all of the MediaWiki “Modern (Grade A)“ browsers. This also includes
polyfills – I’m not sure how okay it is to put them on a CC BY-SA 3.0
page, but hopefully the bottom of the banner comment helps with that.

The build result is not committed, since it looks like there is a chance
it might be determinstic (as long as we commit the package-lock.json and
don’t do anything especially strange in the config). Run `make all` or
`npm run build` to build it – the result is in dist/main.js.

This also means that we can get rid of the custom Array.flatMap polyfill
(core-js ships one that’s undoubtedly better), as well as the header
comment in acdc.js (the build result gets one via the banner plugin).