Include fallback for HTMLInputElement.setRangeText()
This method is not supported on all modern browsers, so we have to
include *some* fallback behavior. There are other options than just
overriding the input value, but I can’t really be bothered to research
them (and *their* compatiblity) – the overwhelmingly common case should
be pasting into an empty input anyways, in which case this is fine.