Add “load category” button (fixes #1)
The implementation is fairly simple, with the “microsleep” copied from
the “load PagePile” feature (I might extract that to a function later).
The tough part was finding a good way to add the button to the UI.

A ListToolGroup would be ideal – in a way, “load category” and “load
PagePile” are tools, and the list in a popup is exactly what we want.
But the tool parts of OOUI seem to be very much a thing of their own,
that we can’t rip any parts out of – so I don’t see how to integrate
this with the ellipsis icon. Perhaps later I’ll update the AC/DC UI so
it can accommodate a proper toolbar – in that case it would probably
also become a “full-page app”, more like VisualEditor.

Visually, making the buttons frameless and the surrounding panel layouts
framed instead also looks good – the problem is that the buttons don’t
actually occupy the full space of the panel, so there’s an area that
looks like it’s part of the button (as it’s within the panel layout’s
frame) but actually isn’t (i. e. it doesn’t get a special cursor and you
can’t click on it). Making the button (<span> and <a>) width: 100% via
CSS almost solves that, but not quite, because the icons seem to involve
a bit of a negative margin that’s then left over as non-button space at
the other end even at 100% width.

So for now, we just place the buttons beneath each other, with no
special tweaks, and make the whole surrounding stack layout padded so
that it doesn’t give the impression that the buttons are supposed fill
it perfectly – for now, it’s not perfect and that’s okay.

(“tag” is the icon that VisualEditor uses for categories, so we use it
for the same purpose here.)