Ensure prompts display above AC/DC dialog
OOUI uses a default window manager for the prompt, alert and confirm
functions, and if our own window manager is inserted in the DOM after
that one, the AC/DC dialog will cover the auxiliary dialogs, whereas we
need them to cover the AC/DC one (for the “load from” features).

The default window manager is created on-demand, so the described
problem mostly occurs during development – load AC/DC, create a prompt,
close and reload AC/DC, and then prompts from there will not be visible.
However, I suppose that this could also happen to regular users if they
cause any other prompts/alerts/confirms to occur before opening AC/DC.

OOUI doesn’t really support multiple windows at once, by the way – if we
tried this in a single WindowManager, it would complain – so perhaps
I’ll eventually overhaul the UI so it’s not a dialog but more of a
full-page app like VisualEditor. For now, though, this works.