Fix detection of existing statements
With WikibaseMediaInfo change I3970236fae (commit 6ca1ca8bb0 [1]), the
propertyId of a StatementWidget moved into a state member. The effect of
this was that, when searching for existing statements, we would always
end up looking for ….statements[ undefined ] and conclude that there
were no existing statements, resulting in duplicate ones.