Remove files from input after they’re done
After we’ve added all statements to a file, remove it from the files
widget. This gives users an alternative indicator of exactly where the
script currently is, in addition to the more general information of the
progress bar. It also means that if the save cancels at some point for
any reason, then the dialog should be immediately ready to retry – all
the files that were already done are gone, all the ones that are left to
do are still there. (An immediate retry may not be a great idea after an
error, but I plan to add a “cancel” feature – in that case, this would
effectively let you suspend and then later resume a batch of edits, as
long as you keep the page open.)

This requires a change to titlesToEntityIds() so that we still have the
titles, not just the entity IDs, in the loop.