Remove workaround for T230160
My fix for T230160 [1] (OO.ui.Process does not support native promises)
was merged and deployed, so we don’t need to work around it anymore.
This effectively reverts commit bcd41ca9ff, except that the `this`
second argument to Process isn’t reinstated – we don’t need it because
we use an arrow function (which preserves the outer `this` anyways).

[1]: https://phabricator.wikimedia.org/T230160