Add “stop” button
This splits the dialog into two modes – “edit” mode is the regular mode
and shows all the buttons/actions we had until now, whereas “save” mode
is entered as soon as the save begins and replaces the “publish” button
with a “stop” button. When that button is pressed, it resets the dialog
to “save” mode and sets a flag which informs the still-running save
process that it should stop without making any more edits.

OOUI doesn’t usually trigger further actions while one is still pending,
so we have to override onActionClick() to make this work. Unfortunately,
custom config elements aren’t preserved by ActionWidget as far as I can
tell, so we have to match on the action name directly. (I assume a magic
name in the flags would also work, but that doesn’t seem better to me.)

The progress bar is still there after the save is stopped (or finished,
for that matter, if you re-open the dialog after it closes itself), but
I’ll fix that separately.