Tag edits with “ACDC”
WikibaseMediaInfo’s StatementWidget gained `summary` and `tags` options,
so now we can tag our edits. To keep the script wiki-agnostic, we only
use the “ACDC” tag on Commons and Test Commons; other admins can adjust
the script as necessary for their wikis (and I’d be happy to merge such
changes into this version as well).

The way we pass the tags into the StatementsDialog is a bit awkward,
unfortunately – the arguments to windowManager.openWindow() are only
passed to getSetupProcess(), not to the constructor (after all, the same
window can be opened multiple times), so we have to copy them there as
well. (We could get rid of the tags support in the constructor config
completely, but I prefer to keep it.)