Add browser test for adding a statement
It’s not too pretty, but it works for now. Most of the browser
interaction should probably be wrapped up into a page object at some
point; the entity reset at the beginning is also more cumbersome that it
needs to be due to a missing WikibaseMediaInfo feature (T233522 [1]).

Since this test actually edits the target wiki, it requires credentials.
It expects to find a MEDIAWIKI_USERNAME and MEDIAWIKI_PASSWORD in the
environment, which should form a bot password [2]. Locally, I have those
saved in a .mw_credentials file (hence the .gitignore change) like this:

    export MEDIAWIKI_USERNAME='...'
    export MEDIAWIKI_PASSWORD='...'

This allows me to run the tests with commands like the following:

    (. .mw_credentials && npm test)
    (. .mw_credentials && npm run wdio)

In Travis CI, the variables are configured in the Travis settings.

[1]: https://phabricator.wikimedia.org/T233522
[2]: https://www.mediawiki.org/wiki/Special:MyLanguage/Manual:Bot_passwords