Fix race condition in injecting AC/DC in browser tests
It turns out that, while the presence of mediaWiki and jQuery in the
window ensures the presence of mw.loader and most of its members, it
does not ensure the presence of mw.loader.using, which is added later,
in the mediawiki.base module. As such, if we don’t explicitly wait for
mw.loader.using to be available as well, it’s possible for AC/DC’s code
to run before mw.loader.using becomes available, leading to an instant
crash and a failed browser test.