Extract page objects for statements browser test
Split into three classes now: StatementsWidget, with each instance
encapsulating an instance (there can be more than one StatementsWidget);
WikibaseMediaInfo, so that entitySelectorEntry can be used from both
ACDC and StatementsWidget without a circular dependency; and ACDC (now
with an uppercase filename to match its class and the other two).

Depending on your Git settings, the move of acdc.js to ACDC.js may be
reported as a deletion and addition instead due to the many changes in
the file; adding -M3 works for me.