Load page IDs in Node.js using MWBot
If we try to load the page IDs in the browser, in a “before” hook, then
it’s impossible to run any of the statement tests by itself because the
“before” hook will only work if another test has run before it
(otherwise there’s no page to provide the mediaWiki global). Instead, we
can directly make the API call in Node.js (using the MWBot library,
which isn’t as nice as mw.Api but at least better than making network
requests directly), which should also be a bit more efficient (cut out
communication with the browser).