Update for ResourceLoader module changes
The wikibase.datamodel.* and wikibase.serialization.* modules were
merged into one (each), so we should import them differently.
(Currently, the old modules are still available in production, but
already removed in beta, where this was noticed by the browser tests.)

To make testing in both environments simpler, I added a commented-out
alternative base URL to wdio.conf.js and split .mw_credentials into
.mw_credentials_beta and .mw_credentials_test, to be used like this:

    (. .mw_credentials_beta && make check)
    (. .mw_credentials_test && make check)

Depending on which line in wdio.conf.js is currently active.

(wikibase.utilities.)ClaimGuidGenerator is also prepared for new-style
imports (as in, assigned to a top-level variable instead of directly
accessing it in wikibase.util where it’s used), though that module
is still not exported that way yet.