Prepare for other datatype support in WikibaseMediaInfo
WikibaseMediaInfo will soon gain support for additional datatypes, which
means we need to pass the correct datatype of the selected property into
the StatementWidget. Fortunately, we already receive this as part of the
arguments to the AddPropertyWidget’s “choose” event – I just hadn’t seen
it before because I was looking in the first argument, which seems to be
some widget (I’m not sure which one, actually), while the data is really
the second argument.

Fixes #3 – thanks a lot to @egardner for the notification!