Fix submit button click in browser tests
WikibaseMediaInfo no longer seems to emit the “change” event on a
StatementWidget as soon as the item has been selected by click, but only
after it has been fully added to the DOM (which involves an API call to
wbformatvalues). This means that we need to wait for the submit button
to become enabled in the browser tests before clicking it, otherwise the
click will have no effect and we’ll time out waiting for the dialog to
close itself.