Add rudimentary WebExtension for development
Firefox removed the Scratchpad in version 72, so instead let’s use a
WebExtension to easily run the code in a development environment. (Since
Firefox 72, the Web Console can also open and save files, making it a
more suitable Scratchpad replacement, but this was completely missing
when the removal of the Scratchpad was announced.)

I first tried to make this work as a user script, a new WebExtensions
API [1] available since Firefox 68 (around the same time that the end of
the Scratchpad was announced). This took quite some time to get working
(as far as I can tell, the documentation on MDN is not merely incomplete
but in parts flat-out wrong; what worked for me was "user_scripts": {}
in manifest.json and defining the source only in register()), but it
doesn’t do the right thing: user scripts get the same clean view of the
DOM as content scripts [2], so the mediaWiki and jQuery globals aren’t
available to acdc.js if it’s loaded that way. As far as I can tell, the
intention of user scripts (and their difference from content scripts) is
not, as I thought, to have scripts with full access to the page’s
JavaScript, but to further isolate them within the WebExtension [3].

So ultimately, I’m exploiting the fact that in Firefox, a content script
can use window.eval() to evaluate code in the page context, whereas
eval() evaluates code in the content script context [4]. This is very
ugly and doesn’t work in any other browser, but as far as I can tell
it’s the only way to make this work.

The "web_accessible_resources" key in manifest.json doesn’t seem to be
necessary, by the way – the fetch() of acdc.js still succeeds without it
– but I’m not sure if this is intentional or a bug (MDN says [5] that
“content scripts don’t need to be listed as web accessible resources”,
but I’m not registering acdc.js as a content script), so let’s just be
on the safe side and list it anyways.

[1]: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/userScripts
[2]: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Content_scripts#DOM_access
[3]: https://wiki.mozilla.org/WebExtensions/UserScripts
[4]: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Content_scripts#Using_eval_in_content_scripts
[5]: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/web_accessible_resources