Add i18n support
Messages may be specified in JSON at [[MediaWiki:Gadget-ACDC-i18n.json]]
([1] on Commons). This is optional – the English messages are still
hard-coded into the gadget as a fallback, in case the JSON page is not
defined. (But if it is defined, the English messages within it can
override the hard-coded ones.)

[1]: https://commons.wikimedia.org/wiki/MediaWiki:Gadget-ACDC-i18n.json