Add i18n for three placeholders
I’m not sure if the “Category:” and “File:” namespaces can be translated
or not – for now I’m instructing translators to leave them alone.

The “12345” placeholder for the PagePile ID is left untranslated, since
it must always look like this (using ASCII Arabic numerals) even in
languages which write numbers differently – the PagePile IDs are the
same, after all.