Fix StatementWidget instantiation in sanity check
In 15f65c9614, I updated the real StatementWidget instantiation in the
StatementsDialog but forgot the one in the sanity check. This worked at
the time because WikibaseMediaInfo included some backwards compatibility
code, but this was more recently removed in WikibaseMediaInfo change
Iaa13136693 (commit 18edbc6950), breaking the sanity check.

While we’re at it, we can also remove some of our own backwards
compatibility code now, since the propertyType support is deployed
everywhere now and the properties map is no longer read anywhere.