Move some devDependencies to dependencies
I previously wasn’t sure what the point of non-dev dependencies for this
gadget should be (because it doesn’t run via npm at all), but I think
there’s actually a split that makes a lot of sense: everything you need
to build the gadget is a prod dependency, everything you need to test it
is a dev dependency. (Here, “test” includes linting, so it’s not just
about unit tests.)

One benefit of this is that, when GitHub warns me about a vulnerability
in a dependency, I can now relatively easily tell whether that affects
Commons (the dependency is in `npm ls --prod`) or not.