Stop using StatementWidget.submit()
Recent WikibaseMediaInfo changes broke AC/DC when adding qualifiers:
after saving the statement, WikibaseMediaInfo would attempt to find the
original statement in its getData() in order to update it, but this
would fail since its getData() would not contain the statement we had
injected in getChanges().

I initially tried to fix this by overriding getData() as well, but I
don’t think that’s a good idea. submit() does a lot of things by now: it
compares the initial with the current state, sends the needed API
requests, and then updates its state accordingly. We don’t want most of
these things – in particular, we certainly don’t the widget to update
itself after the first save, potentially affecting future saves. So it
seems easier to just make the API call ourselves and discard the rest of
submit(), sidestepping the problem.

A side effect is slightly stronger “assertuser” checking. We use the
mw.Api.assertCurrentUser() function that was recently introduced in
MediaWiki, whereas the WikibaseMediaInfo code (which predates it) has
slightly simpler conditions for adding assert/assertuser parameters; one
difference is that assertCurrentUser will cause an error if the page
things we’re anonymous whereas we’re actually logged in, which was not
the case before. Consequently, in the browser tests, we need to inform
MediaWiki that we’ve logged in to avoid errors.