Handle errors better
The Process promise should only reject with OOUI Errors, not arbitrary
objects, or the error handling will itself throw more errors.

Almost all of the save code is extracted into a method (a function local
to the switch/case would not be allowed by the linter and also it would
just be ugly), with just the process/action-specific code remaining
within getActionProcess.