Rename things in preparation for “remove” feature
To accommodate an upcoming feature to remove statements, rename some
things (variables, methods, etc.) by adding a “to add” infix. (In some
cases it acts as a suffix because there’s no “type” like “widget” or
“field” at the end.)

Naming things is hard, but this seems to work better than a “to add”
suffix (statementsFieldToAdd) or an “add” prefix (addAddPropertyWidget).

This also affects some message keys, but so far I’m not aware of anyone
having translated messages, so I’m not too worried about changing that.