Add preliminary support for removing statements
This adds a second section of statements to remove below the one for
statements to add. So far, only “simple” statements (without qualifiers)
can be supported, as there are two possible interpretations for
qualifiers: remove only the qualifiers but leave the statements, or
remove the whole statement only when it has those qualifiers – by
disallowing this case right now, we leave either option open for the
future. Also, like for adding statements, it’s not allowed to have the
same property and value more than once. It also shouldn’t be allowed to
have the same property and value in the statements to add and statements
to remove, but that check isn’t implemented yet.

Because I couldn’t figure out a good user interface for this yet – I
feel like the current section takes up a lot of space and is more
prominent than it should be – for now this is hidden behind a global
variable that the user needs to set in their common.js [1]:

    window.acdcEnableRemoveFeature = true; // temporary

I hope to fix that soon; either I’ll find a better way to present the
option of removing statements, or if I hear from users that it’s good as
it is, I may just make it unconditionally available in its current form.

[1]: https://commons.wikimedia.org/wiki/Special:MyPage/common.js