Use POST to get page IDs for titles
With long image titles (specifically, this issue was discovered on the
files in Category:Rumen Radev in 2018 [1]), it’s possible that the list
of titles is too long to fit in a URL, and we’ll get an HTTP 414 error
“request URI too long”. Using POST avoids this.

It would probably be possible to try the GET first, and only fall back
to POST if necessary. But as far as I’m aware, there is currently little
benefit to this – the API results are not cached (unless we specify
maxage or smaxage ourselves), which would otherwise be the main reason
I’m aware of to prefer GET (this matters for the query service, for
example). Maybe when multi-DC MediaWiki is a thing, and GET requests can
be serviced by a secondary data center but POST requests can’t, it would
make more sense to try GET first and benefit from that.

All the other GET requests should™ be fine, as far as I can tell the
length of their parameters is more limited.

[1]: https://commons.wikimedia.org/wiki/Category:Rumen_Radev_in_2018