Fix CSS selector after WikibaseMediaInfo change
With the introduction of a user interface for somevalue and novalue, the
.oo-ui-menuOptionWidget selector no longer matches the entity selector
suggestions, but the “custom value”, “some value” and “no value”
options. Add a surrounding .oo-ui-lookupElement-menu to the selector to
avoid this (while both the entity and snak type selector have the class
oo-ui-menuSelectWidget, only the former has oo-ui-lookupElement-menu).