BREAKING: Polyfill Promise via ResourceLoader instead
MediaWiki now ships a Promise polyfill which can be loaded as a
ResourceLoader module, but is skipped if the browser already defines
Promise (which is the case for most page loads), saving some bandwidth.
I edited the gadgets-definition to make the ACDC gadget depend on that
ResourceLoader module [1], so now we can instruct webpack to not include
the polyfill for Promise, which shrinks the built file by almost 9%
(from 84965 bytes to 77464 bytes).

Any third-party wikis using this gadget must also add this dependency to
their gadgets-definition, or the gadget will no longer work on IE11 and
a few other rare browsers. (It will still work on modern browsers that
include a native Promise definition.)

[1]: https://commons.wikimedia.org/wiki/Special:Diff/421108984