Add `npm run analyze` command
Adjust the webpack config so that a stats file is written next to
main.js, and add a command to analyze that stats file (opens in the
browser automatically).

This should help me figure out which modules to try removing from the
bundle. For example, if this commit is cherry-picked onto the parent
commit’s parent (i. e., remove “polyfill Promise via ResourceLoader”),
it shows that es.promise is the largest single module inside
node_modules/modules/, confirming it was a good candidate for removal.

Note that `npm run analyze` only launches the analyzer, it doesn’t build
the stats file. You usually want to combine it with a build:

   npm run build && npm run analyze