Update browser test property IDs
Test Commons used to use property and item IDs from real Wikidata, but
it looks like at some point it switched to Test Wikidata, so the browser
tests would fail if run against Test Commons, because P694 is not an
Item property there. Choose two new property IDs: the first, P734, has
the type Item on Beta, Test, and real Wikidata, and should be suitable
for any config; unfortunately I was unable to find a second such
property ID [1], and so the second, P116, merely has the type Item on
Beta and Test Wikidata.

[1]: https://twitter.com/LucasWerkmeistr/status/1266709656036048897