Update PagePile URL for toolforge.org migration
The old URL redirects to the new one, but the redirect isn’t
CORS-enabled, so it’s of little use to us.

(I committed and deployed this on 2020-06-26, but then apparently didn’t
push the main branch to GitHub, so this got lost with the next
deployment when that was done from a different device that didn’t have
the commit. Rebasing and redeploying it now. I haven’t heard any
complaints about PagePile import being broken so maybe the redirect is
CORS-enabled now?)