Add “favorite properties” feature
Users can specify (e. g. in their common.js [1]) the globals
acdcFavoritePropertiesToAdd and acdcFavoritePropertiesToRemove,
whereupon AC/DC will add statement widgets for those properties on
startup, without the user having to add the properties by hand. The
favorite properties to add default to the WikibaseMediaInfo default
properties (i. e.: “depicts”). As a shortcut, acdcFavoriteProperties can
be used to set both lists at once (or, for users with the remove feature
disabled, just as a shorter version of acdcFavoritePropertiesToAdd).

In the browser tests, we need to disable the default favorite properties
because they break some of the tests. Maybe I’ll later write dedicated
tests for this feature. (Honestly, I probably won’t, but it’s one in the
morning and I want to get this feature deployed without broken Travis.)

[1]: https://commons.wikimedia.org/wiki/Special:MyPage/common.css