WIP: Use expect-webdriverio
Start to use expect-webdriverio [1] a bit. Works in `npm run wdio`, but
`npm run lint` complains that the “expect” global isn’t defined, and to
be honest, I have no clue where it comes from, or what the right env to
add to .eslintrc.js would be.

Can perhaps be resurrected as part of a larger effort to move *all*
tests to expect-style instead of assert-style.

[1]: https://www.npmjs.com/package/expect-webdriverio