Check if login in browser tests worked
Currently, login is failing, probably due to problems on the Beta
cluster. Previously, this wasn’t detected, and the tests would fail
later when assert: 'user' failed, which is a bit confusing. Instead,
check that the login worked, and immediately throw an error if it
didn’t, with the message from the API:

> Login failed: Unable to continue login. Your session most likely timed out.