Check if already logged in in browser tests
The browser tests are now failing because we are trying to log in in a
bot password session (i. e., we’re already logged in), and bot password
sessions don’t allow logging in again. I’m not sure why this is new –
did MediaWiki used to allow this, or did something change in the browser
setup that now keeps the session alive for longer? – but it’s relatively
simple to work around: check if we’re already logged in, and if yes,
skip the log in. We can even combine this into the same action=query
call that we need anyways to get the login token.

However, this requires that we know the login username. (In principle,
we could just check if ! 'anon' in response.query.userinfo, but I don’t
feel comfortable with that.) This means that the login form with the
bott password app ID in the user name (User@ACDC) is no longer supported
– the app ID must now be at the beginning of the password. (We could try
to detect the app ID in the user name and split it out, but why bother?
Safer to treat the user name and password as opaque.)

See also https://phabricator.wikimedia.org/T263617.