Switch to GitHub actions
Travis CI now only gives you a fixed allotment of “credits”, and once
you’ve used those up, it’s over. (You can apparently request more and
it’ll be decided on a “case-by-case basis”, but why should I bother?)
This makes it especially important to migrate the daily test runs off
of it, and AC/DC is one of those (the other is Wikidata Lexeme Forms).