Fix early return in FileInputWidget.getLookupRequest()
If we want to effectively return an empty list of search results, we
need to return an object from which getLookupCacheDataFromResponse()
will read an empty array, not directly an empty array. This caused a
sporadic bug [1]: TypeError: response.query is undefined.

In the test, we have to manually register a global error handler and
assert that it doesn’t get triggered – WebdriverIO or Mocha or whatever
don’t do this for us, apparently.

[1]: https://phabricator.wikimedia.org/T279852

Bug: T279852