Install global error handler for all tests
This generalizes the global error handler added in the previous commit
for one test: really, we want to be checking for uncaught client-side
errors in every test.

installGlobalErrorHandler() can’t be a beforeEach() function because the
event listener gets reset on each page navigation, and page navigation
happens as part of a test, not before it. However, we can put it in
injectAcdc(), which is called by every test and is therefore almost as
good. In the afterEach() handler, we add a debugger; statement, so that
a developer can easily inspect the error in the browser’s dev tools.