Hide StatementWidget footer
This footer contains several buttons – “publish changes”, “cancel”, and
“remove all”. The first two do nothing (though “cancel” first shows a
popup asking for confirmation, before proceeding to do nothing if you
confirm it), and the third one removes all the statements from the list
to be added, but won’t actually remove anything from the statements of
any file, so it’s still confusing. In short, we don’t want any part of
this footer. I don’t see a way to disable it via any parameters or
options to the StatementWidget, so let’s just hide it via CSS.