Create edit group for ≥10 statements
EditGroups for Commons have been available for a while; let’s
automatically create a group for batches with ten or more statements
(which is just an arbitrary threshold for now, but I don’t want to
complicate the interface by making it configurable for users if it’s not
necessary).

Implements #15.