Populate files widget with selected Cat-a-lot files
This attempts to, on a best-effort basis, use the currently selected
files in the Cat-a-lot gadget as the files to edit with AC/DC. If the
process fails for any reason (e.g. Cat-a-lot isn’t loaded, or isn’t
currently enabled [its method crashes in that case], or its API [which
has no kind of stability promise as far as I’m aware] changes), then we
silently ignore the error. If someone notices that this feature no
longer works, and reports it, I may or may not fix it, but I can’t make
any guarantees for this integration with someone else’s gadget.

Fixes #12.